requirements/dev.txt
-r base.txt

behave_django==1.4.0
factory_boy==3.0.1
selenium==3.141.0
coverage==5.2.1
flake8==3.8.3
page_objects==1.1.0